pyproject.toml
[tool.poetry]
name = "umaru"
version = "1.0"
description = "A small Switch Pro Controller driver which lets you use your computer with it !"
license = "MIT"
authors = ["Animenosekai <animenosekai.mail@gmail.com>"]
maintainers = ["Animenosekai <animenosekai.mail@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Animenosekai/umaru"
documentation = "https://github.com/Animenosekai/umaru/blob/main/README.md"
keywords = [
    "animenosekai",
    "umaru",
    "console",
    "switch",
    "controller",
    "pro controller",
    "game controller",
    "game",
]
classifiers = [
    # Status
    "Development Status :: 4 - Beta",

    # Environment
    "Environment :: Console",

    # Audience
    "Intended Audience :: End Users/Desktop",

    # Licensing
    "License :: OSI Approved :: MIT License",

    # Software Requirements
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",

    # Topics
    "Topic :: System :: Hardware",

    # Code quality
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.7"
pynput = "^1.7.6"
hidapi = "^0.14.0"
rich = "^13.5.1"

[tool.poetry.scripts]
umaru = 'umaru.__main__:entry'

[tool.poetry.urls]
"Issue Tracker" = "https://github.com/Animenosekai/umaru/issues"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
